Move client package out of pkg/
