+++ client/auth/auth.go
@@ -12,7 +12,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/cozy/cozy-stack/pkg/client/request"
+	"github.com/cozy/cozy-stack/client/request"
 	"github.com/cozy/cozy-stack/pkg/config"
 )
